# spec/openapi.yaml
openapi: 3.1.0
info:
  title: ActNow API
  version: 0.1.0-draft
  description: >
    P0 API contract for ActNow / AIGC-Platform. This file is the machine-readable API
    truth source. Explanatory planning lives in tech/03-api-contract.md.
servers:
  - url: http://localhost:3000/api
    description: local Docker Compose
tags:
  - name: Projects
  - name: Agent
  - name: Canvas
  - name: Assets
  - name: Storyboard
  - name: Generation
  - name: Exports
paths:
  /projects:
    post:
      tags: [Projects]
      summary: Create a demo project and default workspace
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Created project with workspace aggregate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceAggregate'
  /projects/{projectId}/workspace:
    get:
      tags: [Projects]
      summary: Get workspace aggregate for the single-page React Flow workbench
      operationId: getWorkspace
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Workspace aggregate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceAggregate'
  /projects/{projectId}/script/lock:
    post:
      tags: [Projects]
      summary: Lock current script draft and initialize scenes, shots, and canvas
      operationId: lockScript
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [script_version]
              properties:
                script_version:
                  type: integer
      responses:
        '200':
          description: Locked script and updated workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceAggregate'
  /projects/{projectId}/canvas:
    put:
      tags: [Canvas]
      summary: Save React Flow canvas document
      operationId: saveCanvas
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CanvasDocument'
      responses:
        '200':
          description: Saved canvas document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CanvasDocument'
  /agent/threads/{threadId}/messages:
    post:
      tags: [Agent]
      summary: Send user message into ActNow Harness
      operationId: createAgentMessage
      parameters:
        - $ref: '#/components/parameters/ThreadId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentMessageRequest'
      responses:
        '202':
          description: Message accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentAcceptedResponse'
  /agent/threads/{threadId}/events:
    get:
      tags: [Agent]
      summary: Read ActNow Harness event stream
      operationId: listAgentEvents
      parameters:
        - $ref: '#/components/parameters/ThreadId'
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Event batch. SSE can reuse the same event schema.
          content:
            application/json:
              schema:
                type: object
                required: [events]
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/AgentEvent'
                  next_cursor:
                    type: string
                    nullable: true
  /projects/{projectId}/assets:
    post:
      tags: [Assets]
      summary: Create asset metadata after upload or generation
      operationId: createAsset
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssetRequest'
      responses:
        '201':
          description: Created asset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
  /projects/{projectId}/storyboard/generate:
    post:
      tags: [Storyboard]
      summary: Generate or regenerate storyboard shots from script/context
      operationId: generateStoryboard
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                episode_id:
                  type: string
                scope:
                  type: string
                  enum: [episode, scene, selected_shots]
                target_ids:
                  type: array
                  items:
                    type: string
      responses:
        '202':
          description: Accepted as task or tool call
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskAcceptedResponse'
  /shots/{shotId}:
    patch:
      tags: [Storyboard]
      summary: Update one shot
      operationId: updateShot
      parameters:
        - name: shotId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateShotRequest'
      responses:
        '200':
          description: Updated shot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shot'
  /generation-tasks:
    post:
      tags: [Generation]
      summary: Create generation task using BullMQ + Redis worker queue
      operationId: createGenerationTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGenerationTaskRequest'
      responses:
        '201':
          description: Created generation task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerationTask'
  /generation-tasks/{taskId}:
    get:
      tags: [Generation]
      summary: Get generation task state from PostgreSQL truth source
      operationId: getGenerationTask
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Generation task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerationTask'
  /generation-tasks/{taskId}/prompt-package:
    get:
      tags: [Generation]
      summary: Get prompt package for human-loop generation
      operationId: getPromptPackage
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Prompt package
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromptPackage'
  /generation-tasks/{taskId}/uploads:
    post:
      tags: [Generation]
      summary: Upload human-loop generation result
      operationId: uploadGenerationResult
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                manifest:
                  type: string
                  description: Optional JSON manifest mapping task_id to files.
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: Uploaded and validated files
          content:
            application/json:
              schema:
                type: object
                required: [files, task]
                properties:
                  files:
                    type: array
                    items:
                      $ref: '#/components/schemas/GeneratedFile'
                  task:
                    $ref: '#/components/schemas/GenerationTask'
  /generation-tasks/{taskId}/retry:
    post:
      tags: [Generation]
      summary: Retry generation task
      operationId: retryGenerationTask
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Retried task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerationTask'
  /generation-tasks/{taskId}/cancel:
    post:
      tags: [Generation]
      summary: Cancel generation task
      operationId: cancelGenerationTask
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Canceled task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerationTask'
  /exports:
    post:
      tags: [Exports]
      summary: Create composition/export job
      operationId: createExportJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project_id]
              properties:
                project_id:
                  type: string
                episode_id:
                  type: string
      responses:
        '202':
          description: Export job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskAcceptedResponse'
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
    ThreadId:
      name: threadId
      in: path
      required: true
      schema:
        type: string
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
  schemas:
    CreateProjectRequest:
      type: object
      required: [initial_input]
      properties:
        title:
          type: string
        route:
          type: string
          enum: [short_drama, comic_drama]
          default: short_drama
        initial_input:
          type: string
    Project:
      type: object
      required: [id, title, route, current_stage, status]
      properties:
        id: { type: string }
        title: { type: string }
        route: { type: string }
        current_stage: { type: string }
        status: { type: string }
        settings:
          type: object
          additionalProperties: true
    WorkspaceAggregate:
      type: object
      required: [project, canvas, agent_thread]
      properties:
        project:
          $ref: '#/components/schemas/Project'
        canvas:
          $ref: '#/components/schemas/CanvasDocument'
        agent_thread:
          $ref: '#/components/schemas/AgentThread'
        scenes:
          type: array
          items:
            $ref: '#/components/schemas/Scene'
        shots:
          type: array
          items:
            $ref: '#/components/schemas/Shot'
        assets:
          type: array
          items:
            $ref: '#/components/schemas/Asset'
        generation_tasks:
          type: array
          items:
            $ref: '#/components/schemas/GenerationTask'
    CanvasDocument:
      type: object
      required: [project_id, nodes, edges, viewport, version]
      properties:
        project_id: { type: string }
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/CanvasNode'
        edges:
          type: array
          items:
            type: object
            additionalProperties: true
        viewport:
          type: object
          additionalProperties: true
        version:
          type: integer
    CanvasNode:
      type: object
      required: [id, type, position, data]
      properties:
        id: { type: string }
        type: { type: string }
        position:
          type: object
          required: [x, y]
          properties:
            x: { type: number }
            y: { type: number }
        data:
          type: object
          additionalProperties: true
    AgentThread:
      type: object
      required: [id, project_id, status]
      properties:
        id: { type: string }
        project_id: { type: string }
        status: { type: string }
        focus_ref:
          $ref: '#/components/schemas/TargetRef'
    CreateAgentMessageRequest:
      type: object
      required: [content]
      properties:
        content: { type: string }
        focus_ref:
          $ref: '#/components/schemas/TargetRef'
        client_context:
          type: object
          additionalProperties: true
    AgentAcceptedResponse:
      type: object
      required: [thread_id, message_id]
      properties:
        thread_id: { type: string }
        message_id: { type: string }
    AgentEvent:
      type: object
      required: [id, thread_id, event_type, actor, created_at]
      properties:
        id: { type: string }
        thread_id: { type: string }
        task_id: { type: string, nullable: true }
        event_type: { type: string }
        actor: { type: string, enum: [human, agent, system, worker] }
        payload:
          type: object
          additionalProperties: true
        created_at: { type: string, format: date-time }
    Scene:
      type: object
      required: [id, episode_id, title, order]
      properties:
        id: { type: string }
        episode_id: { type: string }
        title: { type: string }
        order: { type: integer }
    Shot:
      type: object
      required: [id, scene_id, order, description, status]
      properties:
        id: { type: string }
        scene_id: { type: string }
        order: { type: integer }
        description: { type: string }
        camera:
          type: object
          additionalProperties: true
        emotion: { type: string }
        duration: { type: number }
        status: { type: string }
    UpdateShotRequest:
      type: object
      properties:
        description: { type: string }
        camera:
          type: object
          additionalProperties: true
        emotion: { type: string }
        duration: { type: number }
        version: { type: integer }
    CreateAssetRequest:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [character, character_form, scene, prop, reference, panorama, simple_card]
        name: { type: string }
        uri: { type: string }
        metadata:
          type: object
          additionalProperties: true
    Asset:
      type: object
      required: [id, type, status]
      properties:
        id: { type: string }
        type: { type: string }
        name: { type: string }
        uri: { type: string }
        status: { type: string }
        metadata:
          type: object
          additionalProperties: true
    CreateGenerationTaskRequest:
      type: object
      required: [project_id, target, gen_type, backend_mode]
      properties:
        project_id: { type: string }
        target:
          $ref: '#/components/schemas/TargetRef'
        gen_type:
          type: string
          enum: [text, asset_extract, panorama, keyframe, video, export]
        backend_mode:
          type: string
          enum: [human_loop, real_api, sample]
        options:
          type: object
          additionalProperties: true
    GenerationTask:
      type: object
      required: [id, project_id, target, gen_type, backend_mode, status]
      properties:
        id: { type: string }
        project_id: { type: string }
        target:
          $ref: '#/components/schemas/TargetRef'
        gen_type: { type: string }
        backend_mode: { type: string }
        status:
          type: string
          enum: [pending, prompt_ready, waiting_upload, generating, retrying, completed, failed, canceled]
        prompt_package_id: { type: string, nullable: true }
        error_code: { type: string, nullable: true }
    PromptPackage:
      type: object
      required: [id, task_id, prompt, manifest]
      properties:
        id: { type: string }
        task_id: { type: string }
        prompt: { type: string }
        references:
          type: array
          items:
            $ref: '#/components/schemas/Asset'
        parameters:
          type: object
          additionalProperties: true
        manifest:
          type: object
          additionalProperties: true
    GeneratedFile:
      type: object
      required: [id, task_id, file_type, uri, version]
      properties:
        id: { type: string }
        task_id: { type: string }
        file_type: { type: string }
        uri: { type: string }
        version: { type: integer }
        model_meta:
          type: object
          additionalProperties: true
    TargetRef:
      type: object
      required: [type, id]
      properties:
        type:
          type: string
          enum: [project, episode, scene, shot, asset, canvas_node, generation_task]
        id:
          type: string
    TaskAcceptedResponse:
      type: object
      required: [task_id, status]
      properties:
        task_id: { type: string }
        status: { type: string }